:truck: move ServiceUrlField into 'models' folder
